integração api python - dotnet

+++ TeamsReportDashboard/TeamsReportDashboard/Services/Requester/Delete/DeleteRequesterService.cs
@@ -1,4 +1,5 @@
-﻿using TeamsReportDashboard.Interfaces;
+﻿using TeamsReportDashboard.Backend.Exceptions;
+using TeamsReportDashboard.Interfaces;
 
 namespace TeamsReportDashboard.Backend.Services.Requester.Delete;
 
@@ -18,6 +19,15 @@ public class DeleteRequesterService : IDeleteRequesterService
         {
             throw new KeyNotFoundException("Solicitante não encontrado.");
         }
+        
+        var hasReports = await _unitOfWork.ReportRepository.HasReportsForRequesterAsync(id);
+        if (hasReports)
+        {
+            // Lança a sua nova exceção customizada
+            throw new ConflictException("Este solicitante não pode ser excluído pois possui relatórios associados.");
+        }
+
+
         await _unitOfWork.RequesterRepository.DeleteRequesterAsync(id);
         await _unitOfWork.CommitAsync();
     }
